Dernière modif (pitié oui) des classes, instantiations des premiers objet dans le commande runner

+++ src/main/java/com/example/demo/models/Member.java
@@ -22,9 +22,9 @@ public class Member {
     private String firstName;
     private String lastName;
     private String email;
-    private int pswd;
+    private Long pswd;
 
-    public Member(String firstName, String lastName, String email, int pswd){
+    public Member(String firstName, String lastName, String email, Long pswd){
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
